requirements.txt
streamlit
chromadb
sentence-transformers
google-generativeai
beautifulsoup4
requests
